Change position constant type to double

+++ src/main/java/frc/robot/settings/Constants.java
@@ -33,29 +33,29 @@ public final class Constants {
              * and output to the bottom row.
              * Measured in degrees.
              */
-            public static final int LOW_ARM_POSITION = 0;
+            public static final double LOW_ARM_POSITION = 0;
             /**
              * Medium arm position, for output
              * to the middle row. Measured in degrees.
              */
-            public static final int MED_ARM_POSITION = 0;
+            public static final double MED_ARM_POSITION = 0;
             /**
              * Highest arm position, for output
              * to the top row. Measured in degrees.
              */
-            public static final int HIGH_ARM_POSITION = 0;
+            public static final double HIGH_ARM_POSITION = 0;
 
-            
+
             /**
              * Inner wrist position, for storage
              * and protection. Measured in degrees.
              */
-            public static final int COLLAPSED_WRIST = 0;
+            public static final double COLLAPSED_WRIST = 0;
             /** 
              * Outer wrist position, for intake
              * and output. Measured in degrees.
             */
-            public static final int EXTENDED_WRIST = 0;
+            public static final double EXTENDED_WRIST = 0;
 
         }
